requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9
python-dotenv==1.0.0
SQLAlchemy==2.0.23
gunicorn==21.2.0
boto3==1.34.0
psycopg2-binary==2.9.9